Отображает данные юзера после перезагрузки
