+++ project/src/types/auth-data.ts
@@ -1,3 +1,10 @@
+import { UserData } from './user-data';
+
+export type AuthorizedUser = UserData & {
+  email: string;
+  token: string;
+};
+
 
 export type AuthData = {
   email: string;
